Increase the max buffer size for Stayopen
The default max token size for bufio.Scanner is bufio.MaxScanTokenSize.
By default this 1024*64.  This change increases it to 4MB which should
hopefully be big enough to handle huge metadata dumps.

Large metadata usually happens due to a `-binary` flag used to extract
preview images or thumbnails.  If 4MB is not enough its probably better
to a) extract binary data directly or b) rewrite this library :(